pyproject.toml
[project]
name = "galaxy-classification-guide"
version = "0.1.0"
description = "DATA7901 multi-modal galaxy morphology classification (SDSS DR19)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astropy>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "xgboost>=3.2.0",
]
